// src/types/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type Error = {
  __typename?: 'Error';
  error: Scalars['String']['output'];
  message: Scalars['String']['output'];
  statusCode: Scalars['Int']['output'];
};
// ...
export type Location = {
  __typename?: 'Location';
  address?: Maybe<Scalars['String']['output']>;
  alias?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  managingOrganization?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  npi?: Maybe<Scalars['String']['output']>;
  partOf?: Maybe<Scalars['String']['output']>;
  status?: Maybe<Scalars['String']['output']>;
  tag?: Maybe<Scalars['String']['output']>;
  taxId?: Maybe<Scalars['String']['output']>;
  telecom?: Maybe<Array<Maybe<LocationTelecom>>>;
  tenant?: Maybe<Scalars['String']['output']>;
  type?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['Float']['output']>;
};

export type LocationCommandResponse = {
  __typename?: 'LocationCommandResponse';
  resourceID: Scalars['String']['output'];
};
// ...
export type LocationQueryReadResponse = {
  __typename?: 'LocationQueryReadResponse';
  id: Scalars['String']['output'];
  resource: Location;
};

export type LocationTelecom = {
  __typename?: 'LocationTelecom';
  rank?: Maybe<Scalars['Float']['output']>;
  system?: Maybe<Scalars['String']['output']>;
  use?: Maybe<Scalars['String']['output']>;
  value?: Maybe<Scalars['String']['output']>;
};
// ...
export type LocationRemoveMutation = { __typename?: 'Mutation', locationRemove?: { __typename?: 'Error', error: string, message: string } | { __typename?: 'LocationCommandResponse', resourceID: string } | null };

export type LocationPatchMutationVariables = Exact<{
  locationPatchId: Scalars['String']['input'];
  requestBody: LocationPatchInput;
  tenant: Scalars['String']['input'];
}>;


export type LocationPatchMutation = { __typename?: 'Mutation', locationPatch?: { __typename?: 'Error', error: string } | { __typename?: 'LocationCommandResponse', resourceID: string } | null };
// ...
export type LocationReadQuery = { __typename?: 'Query', locationRead?: { __typename?: 'LocationQueryReadResponse', resource: { __typename?: 'Location', address?: string | null, name: string, id: string, status?: string | null, updatedAt?: number | null, type?: string | null, description?: string | null } } | null };